test(vcs): isolate stored token internal coverage
